Added end points (not yet tested will make changes if it doesn't work correctly)
Added EndPoint script where the user can set all the end points and optionally the enemies that are allowed to these end points, the script will then save a dictionary of available end points for every enemy in the set waves for easy access.
Added EndPointEnum.
Added AddEndPointComponents to WaveConfigurator and also to the WaveConfiguratorEditor scripts.
Changed custom classes to structs instead.
Added SetEndPoint to the ISpawner interface.
Implemented setting the end point in the Spawner class.