Fixed endpoints (object doesn't delete when arriving gotta check why maybe implement event instead of deleting)
Moved old scene into new folder.
Checked if endpoints worked (it didn't).
Removed EventListener from the enemy prefabs.
The EndPoint script gets the prefabs object at start, which means that when you instantiate that object it doesn't work anymore. A workaround for this in the spawner script to both input the prefab and instance. Then put the SetAgentDestination script and calls for that script on the instance and get the list of available endpoints from the prefab object.