Changes and groups
Change ProgressiveWithDeployments to Progressive. The idea is that each spawner type will both have the deployments and groups built in. This way I don't have to write 3 different script that are almost identical.
Added PGD#2 RTS script I wrote for getting the units in formation. Cleaned it up a bit and now using it for spawning the groups.

Notes:
- Even though the enemies spawn in groups now they still get their own end point. So this means that they could go to different end point even thought they are in the same group. Will fix this next up.