Added the SemiRandom spawner
SemiRandom spawner still does a deployment per enemy but instead of going in order it takes a random enemy until are enemies have been deployed.
Fixed bug where group counter would continue for the next enemy and so delaying the spawn.
Fixed a bug where the last deployment wouldn't exit the coroutine when the last enemy spawned but waited for a time that it takes for the next deployment.