Small but needed changes
Fixed bug with WaveConfigurator not serializing data (apparently if you do so.ApplyModifiedProperties() after the buttons it doesn't work anymore).
Changed positions of shapes to Transform instead, this should be easier to work with than setting the xyz values of a Vector3.
Fixed Gizmos trying to draw start/endpoints even though the point may be empty (lots of errors).