New spawner added
Added auto deploy options for the spawners.
Renamed variables in PerWaveCustom to be more descriptive.
Moved editor control booleans to a new header in the spawners.
Added logic to the PerWaveRandomPool and verified it working.
Duplicated the WaveConfigurator in the scene to make the PerWaveRandomPool without having to delete the PerWaveCustom.
Disabled V-Sync in the quality settings.