Added TimedSpawner
Added TimedSpawner script to WaveConfigurator and the enums as well.
Added TimedSpawner which uses a single list of enemies. Each enemy has a GameObject, group size and a weight variable. The weight is relative to the other weight values (so weight of 1 vs 1 is 50% just like a weight of 50 vs 50).