Some bigger changes
Removed all ContextMenu's.
Removed Event_AgentSpawnedSetEndPoint.
Added custom editor for WaveManager.
Renamed eventWaveInProgress to eventSpawnerFinished.
Added StopCoroutine if the wave is still in progress and is started again.
Changed finished wave workflow. When the eventSpawnerFinished raises, the WaveManager looks through all the spawner scripts it watches over and checks for their states. If a spawner is done a counter goes up. Every time a counter goes up the counter gets checked against the list size and if it matches a bool gets switched. A demo script that is attached to the same object as the WaveManager has 2 listener methods.